Re-license with Apache-2.0 (no more attribution requirement).
After discussion, we have elected to re-license this source as
Apache-2.0. This change was done in order to remove the
attribution requirement. It is no longer necessary to replicate
this copyright/license inforamtion in debian/copyright (or
similiar) files.

Anyone is permitted to use/fork the previous version with
the previous license. Please note, it would require the original
license/copyright information/requirements to remain in place.